+++ textwire/example/main.go
@@ -14,14 +14,8 @@ var tpl *textwire.Template
 func main() {
 	var err error
 
-	err = textwire.RegisterStrFunc("reverse", func(s string, args ...any) string {
-		runes := []rune(s)
-
-		for i, j := 0, len(runes)-1; i < j; i, j = i+1, j-1 {
-			runes[i], runes[j] = runes[j], runes[i]
-		}
-
-		return string(runes)
+	err = textwire.RegisterStrFunc("_isCool", func(s string, args ...any) any {
+		return s == "John Wick"
 	})
 	if err != nil {
 		log.Fatal(err)